% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid_df.R
\name{vid_merge}
\alias{vid_merge}
\title{Merge .gpx and Objects}
\usage{
vid_merge()
}
\description{
Merges object count dataframe produced by vid_yolo() to the .gpx produced by vid_gpx()
}
\examples{
vid_merge()
}
