% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid_setup.R
\name{vid_setup}
\alias{vid_setup}
\title{Setup command for package}
\usage{
vid_setup(video_location)
}
\arguments{
\item{video_location}{The directory and name of the video file including extension}
}
\description{
Sets working directory, adds output folders, downloads gpx.fmt and yolov3 weights file
}
\examples{
vid_setup("C:/videos/video.mp4");
}
