% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid_split.R
\name{vid_split}
\alias{vid_split}
\title{Video Frame Extraction}
\usage{
vid_split(vid_fps)
}
\arguments{
\item{vid_fps}{The desired rate of frame extraction from the video, in frames per second.}
}
\description{
Sets working directory, adds output folders, downloads gpx.fmt and yolov3 weights file
}
\examples{
vid_split(1)
}
