% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid_yolo.R
\name{vid_yolo}
\alias{vid_yolo}
\title{Process Extracted Frames with YOLOv3}
\usage{
vid_yolo()
}
\description{
Initializes YOLOv3, imports weights file, and runs YOLOv3 on extracted video frames. No arguments.
}
\examples{
vid_yolo()
}
